requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
diffusers==0.31.0
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
grpcio==1.63.2
grpcio-tools==1.63.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.15.0
httpx==0.24.0
huggingface-hub==0.26.2
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.0
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
openai==0.28.0
opencv-python-headless==4.10.0.84
packaging==24.1
param==2.1.1
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
python-telegram-bot==20.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.5
sniffio==1.3.1
stability-sdk==0.8.6
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.66.6
transformers==4.46.2
translate==3.6.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
zipp==3.20.2
